+++ zig/src/day05_2.zig
@@ -23,7 +23,7 @@ const std = @import("std");
 /// number 82, which corresponds to soil 84, fertilizer 84, water 84, light 77,
 /// temperature 45, humidity 46, and location 46. So, the lowest location number
 /// is 46.
-pub fn solve(alloc: std.mem.Allocator, file: std.fs.File) !u32 {
+pub fn solve(alloc: std.mem.Allocator, file: std.fs.File) !u64 {
     var reader = file.reader().any();
     var buffer: [1024]u8 = undefined;
